[ALPS05084703] m4u: coverity[10453428] add negative check for array index
add negative check for array index

MTK-Commit-Id: 535c47bd0764b8c872af9807d13ef1531510a165

CR-Id: ALPS05084703
Feature: [Module]ION/M4U

+++ drivers/misc/mediatek/m4u/mt6779/m4u_hw.c
@@ -1013,7 +1013,7 @@ int larb_clock_on(int larb, bool config_mtcmos)
 #ifdef CONFIG_MTK_SMI_EXT
 	int ret = -1;
 
-	if (larb < ARRAY_SIZE(smi_clk_name))
+	if ((unsigned int)larb < ARRAY_SIZE(smi_clk_name))
 		ret =
 		    smi_bus_prepare_enable(larb, smi_clk_name[larb]);
 	if (ret != 0)
@@ -1029,7 +1029,7 @@ int larb_clock_off(int larb, bool config_mtcmos)
 #ifdef CONFIG_MTK_SMI_EXT
 	int ret = -1;
 
-	if (larb < ARRAY_SIZE(smi_clk_name))
+	if ((unsigned int)larb < ARRAY_SIZE(smi_clk_name))
 		ret =
 		    smi_bus_disable_unprepare(larb, smi_clk_name[larb]);
 	if (ret != 0)
@@ -1897,7 +1897,7 @@ int m4u_reclaim_notify(int port, unsigned int mva, unsigned int size)
 
 int m4u_register_fault_callback(int port, m4u_fault_callback_t *fn, void *data)
 {
-	if (port >= M4U_PORT_UNKNOWN) {
+	if ((unsigned int)port >= (unsigned int)M4U_PORT_UNKNOWN) {
 		M4UMSG("%s fail, port=%d\n", __func__, port);
 		return -1;
 	}
@@ -1908,7 +1908,7 @@ int m4u_register_fault_callback(int port, m4u_fault_callback_t *fn, void *data)
 
 int m4u_unregister_fault_callback(int port)
 {
-	if (port >= M4U_PORT_UNKNOWN) {
+	if ((unsigned int)port >= (unsigned int)M4U_PORT_UNKNOWN) {
 		M4UMSG("%s fail, port=%d\n", __func__, port);
 		return -1;
 	}
